Added option to skip Access Point
Added a command line parameter "-w" which will
disable the installation of a Wireless access
point since that assumes that you have a hard
wired ethernet available that you will route
the access point traffic to.  Also added git
and git-cola to the optional packages list.
Finally if no wireless access point and no
optional packages, the script warns that it
has nothing to do.